changé, les ints en float et rentré des valeurs d'equilibrage

+++ SlidingTower/Assets/Script/Turret/Turret.cs
@@ -13,9 +13,9 @@ public class Turret : MonoBehaviour
     public int numMaxTargets = 5;
 
     [Header ("Value of stat upgrade")]
-    public int[] fireRateBonus;
+    public float[] fireRateBonus;
     public int[] damageBonus;
-    public int[] rangeBonus;
+    public float[] rangeBonus;
 
     private float actualFireRate;
     private float actualDamage;
